requirements.txt
numpy
pandas
PyYAML
requests
torch
